Change of type

+++ Final project/BroadMPIUTWavefront.cpp
@@ -5,9 +5,9 @@
 
 using namespace std;
 
-void init_matrix(double *M, int N) {
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N; j++) {
+void init_matrix(double *M, uint64_t N) {
+    for (uint64_t i = 0; i < N; i++) {
+        for (uint64_t j = 0; j < N; j++) {
             if (j == i)
                 M[i * N + j] = (i + 1) / static_cast<double>(N);
             else
